Выполнение ДЗ "5.14. Модуляция"
