+++ js/data.js
@@ -1,0 +1,31 @@
+const NAMES = [
+  'Дмитрий',
+  'Иван',
+  'Рудольф',
+  'Альф',
+  'Юрий',
+  'Михаил',
+  'Инокентий',
+  'Игорь',
+];
+
+const DESCRIPTIONS = [
+  'Скользи со мной в наше общее будущее.',
+  'Про раннюю осень и запах дымка.',
+  'Или как надули матрас и плавали на нём по озеру.',
+  'Зимнее солнце, собаченьки и мандарины.',
+  'В поиске интересных ракурсов.',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const SIMILAR_COMMENTS_COUNT = 25;
+
+export {NAMES, DESCRIPTIONS, MESSAGES, SIMILAR_COMMENTS_COUNT};
